cqfd: allow cqfd to run without a dockerfile
When using a prebuilt distant docker image, cqfd run will now work when
used with no local dockerfile.

The variable `custom_img_name` had a confusing role, because it was used
both for tagging an image and for pulling a distant image. This commit
makes the role clear for the variable, which is now named
`docker_run_image` and which is clearly used for pulling a distant
image.

Having a way to give an alternate name for the tag of the image
generated by cqfd init has no clear use case